Makes voting process be assynchronous
Was added an ajax call to the same old voting endpoint. To do it was
necessary to change the current view.
Now is not used anymore login_required decorator, but is checked if
request.user is authenticated. If user is not authenticated, a HttpResponse
with status 405 (not-allowed) is returned. The response content is filled with
login url, so the ajax response callback can redirect user to do login.
If user is authenticated and for some reason cannot vote, 405 status
code is returned too and the current page is reloaded.