Providing automatic user language by browser request
For that, I made the following changes to `setting.py`:

- Added LocaleMiddleware to the list of middlewares
- Updated language code pt_BR to IETF-equivalent pt-BR

LocaleMiddleware is the responsible to provide this mechanism. It uses
the "Accept-Language" header values to decide what language to display
to the user.

About the second item, the Django auto discover mechanism expects the
language code to be formatted as defined by IETF.